6. Сортировка | Извлечение таблицы и рефакторинг
