+++ src/app/components/table.jsx
@@ -3,20 +3,24 @@ import TableHeader from "./tableHeader";
 import TableBody from "./tableBody";
 import PropTypes from "prop-types";
 
-function Table({onSort, selectedSort, columns, data}) {
+function Table({onSort, selectedSort, columns, data, children}) {
     return (
         <table className="table">
-            <TableHeader {...{onSort, selectedSort, columns}}/>
-            <TableBody {...{columns, data}}/>
+            {children ||
+                <>
+                    <TableHeader {...{onSort, selectedSort, columns}}/>
+                    <TableBody {...{columns, data}}/>
+                </>}
         </table>
     );
 }
 
 Table.propTypes = {
     onSort: PropTypes.func,
-    selectedSort: PropTypes.object.isRequired,
-    columns: PropTypes.object.isRequired,
-    data: PropTypes.array.isRequired,
+    selectedSort: PropTypes.object,
+    columns: PropTypes.object,
+    data: PropTypes.array,
+    children: PropTypes.array,
 }
 
 export default Table;
