+++ src/app/components/users.jsx
@@ -9,12 +9,34 @@ import UsersTable from "./usersTable";
 import _ from "lodash"
 
 
-const Users = ({users, ...rest}) => {
+const Users = () => {
     const pageSize = 8
     const [currentPage, setCurrentPage] = useState(1)
     const [professions, setProfessions] = useState();
     const [selectedProf, setSelectedProf] = useState();
     const [sortBy, setSortBy] = useState({iter: "name", order: "asc"});
+    const [users, setUsers] = useState()
+
+    useEffect(() => {
+        api.users.fetchAll().then((data) => setUsers(data))
+    }, [])
+
+
+    const handleDelete = (userId) => {
+        setUsers(users.filter(user => user._id !== userId))
+    }
+
+    const handleToggleBookmark = (id) => {
+        const newArr = users.map(i => {
+                if (i._id === id) {
+                    i.bookmark = i.bookmark === false ? true : false
+                }
+                return i
+            }
+        )
+        setUsers(newArr)
+    }
+
     const handlePageChange = (pageIndex) => {
         setCurrentPage(pageIndex)
     }
@@ -36,54 +58,58 @@ const Users = ({users, ...rest}) => {
         setSortBy(item)
     }
 
-    const filteredUsers = selectedProf // && selectedProf._id
-        ? users.filter(user => JSON.stringify(user.profession) === JSON.stringify(selectedProf))
-        : users
-    const count = filteredUsers.length
-    const sortedUsers = _.orderBy(filteredUsers, [sortBy.path], [sortBy.order])
-    const userCrop = paginate(sortedUsers, currentPage, pageSize)
-    const clearFilter = () => {
-        setSelectedProf()
-    }
-    return (
-        <>
-            <div className="d-flex">
-                {professions && (
-                    <div className="d-flex flex-column flex-shrink-0 p-3">
-                        <GroupList
-                            items={professions}
-                            selectedItem={selectedProf}
-                            onItemSelect={handleProfessionSelect}
+    if (users) {
+        const filteredUsers = selectedProf // && selectedProf._id
+            ? users.filter(user => JSON.stringify(user.profession) === JSON.stringify(selectedProf))
+            : users
+        const count = filteredUsers.length
+        const sortedUsers = _.orderBy(filteredUsers, [sortBy.path], [sortBy.order])
+        const userCrop = paginate(sortedUsers, currentPage, pageSize)
+        const clearFilter = () => {
+            setSelectedProf()
+        }
+        return (
+            <>
+                <div className="d-flex">
+                    {professions && (
+                        <div className="d-flex flex-column flex-shrink-0 p-3">
+                            <GroupList
+                                items={professions}
+                                selectedItem={selectedProf}
+                                onItemSelect={handleProfessionSelect}
+                            />
+                            <button
+                                className={"btn btn-secondary mt-2"}
+                                onClick={clearFilter}>Очистить
+                            </button>
+                        </div>
+                    )}
+
+                    <div className="d-flex flex-column">
+                        {<SearchStatus length={count}/>}
+
+                        <UsersTable
+                            users={userCrop}
+                            onSort={handleSort}
+                            selectedSort={sortBy}
+                            onDelete={handleDelete}
+                            onToggleBookmark={handleToggleBookmark}
                         />
-                        <button
-                            className={"btn btn-secondary mt-2"}
-                            onClick={clearFilter}>Очистить
-                        </button>
-                    </div>
-                )}
 
-                <div className="d-flex flex-column">
-                    {<SearchStatus length={count}/>}
-
-                    <UsersTable
-                        users={userCrop}
-                        {...rest}
-                        onSort={handleSort}
-                        selectedSort={sortBy}
-                    />
-
-                    <div className="d-flex justify-content-center">
-                        <Pagination
-                            itemsCount={count}
-                            pageSize={pageSize}
-                            currentPage={currentPage}
-                            onPageChange={handlePageChange}
-                        />
+                        <div className="d-flex justify-content-center">
+                            <Pagination
+                                itemsCount={count}
+                                pageSize={pageSize}
+                                currentPage={currentPage}
+                                onPageChange={handlePageChange}
+                            />
+                        </div>
                     </div>
                 </div>
-            </div>
-        </>
-    )
+            </>
+        )
+    }
+    return 'loading ...'
 }
 
 Pagination.propTypes = {
